Deleted NsdHelper class. Moved it to to ChangeIpDialog
