+++ Hestia/app/src/main/java/hestia/UI/HestiaApplication.java
@@ -4,16 +4,16 @@ import android.app.Application;
 import android.content.Context;
 
 public class HestiaApplication extends Application {
-    private static Context mContext;
+    private static Context context;
 
     @Override
     public void onCreate() {
         super.onCreate();
-        mContext = this;
+        context = this;
     }
 
     public static Context getContext(){
-        return mContext;
+        return context;
     }
